changed docx to html, continued writing proposal
I changed to html in hopes that github will display it when you click on
it. If not, I'll try somethinge else...

I added an introduction and more subsections into "Review of Other
Works," as well as expanding the Apache Server section. I'm hoping to be done with the Review by the end of tomorrow.