Removed net.{c,h}, added proposal.txt
Removed net code because the project plan changed to use an open-source
library at the beginning of the project and then write the networking
myself after the other subprojects (Butler, Attendant) have become more
defined.

Added proposal.txt as txt is a format that I KNOW github will display
in-browser.