worked on proposal, added odt format
I mostly moved a bunch of paragraphs around and worded them better. I
have a better idea of what belongs in each section now. Next commit will
hopefully finish the proposal.

I added the odt format so that it could be downloaded and viewed in a
nicer format than txt, but is still accessible as a txt to view on
Github.