Update configure, makefile, and plan
The plan was updated with dependency information. It was also told to
run butler.guile in order to run the service. Additionally, do_download
was changed to clone from the directory the plan is located in instead
of donwloading from github every time.

The makefile was tidied up. The BUTLER_LOC variable was added and
install now poperly depends on ${BUTLER_LOC}/bulter.guile; all no longer
does.

the sed expression in configure was update to reflect the makefile
change