Updated build files; made butler.guile executable
butler.guile:
added the shebang for guile at the top of the file
so that it can be manually run from the installation
directory (/hab/pkgs/saffronsnail/*); note that this
is not a good long-term solution as habitat is not
aware of the guile runtime dependency.

configure:
eliminated as a result to changes in plan.sh

makefile:
the makefile now uses variables which are set by sed
expressions in thep plan; additionally, it now uses
the install program to install butler.guile (instead
of cp).

plan.sh:
Added the two necesarry configuration steps to
do_build; the separate configure script has been
eliminated as a result.