The send and receive test passes now!
.:
Added debugging info to the cmake flags

source/butler:
Refactored networking.c. create_datagram_socket now takes a bool
indicating whether or not the caller is requesting a listenting socket,
as this is a cleaner interface than requiring the user know how to set
ai_flags properly. Additionally, the file was modified to take into
account the new PORT typedef.

networking.h now contains a typedef for ports (typedef unsigned PORT)

source/butler/test:
Added log.cpp to the list of files that butler-test-suite depends on

modified log to use fprintf(stderr, ...) instead of std::cerr. This is
because I was getting errors when the networking library tried to print
error messages while the gtest function was logging.

removed start_listener_in_new_thread, as I was worried that I was
sending data on the stack back to the test. Upon further reflection this
was certainly not happening, but having that code in a separate function
wasn't really useful and it's working right now so it will stay as it
is.