Continued entrance protocol tests; added headers
I removed some of the test functions in the entrance protocol tests
because I realized that they were testing implementation and not
requirements.

Added the entrance-{request,response}.h files so that modules that need
to see the internals can. Note that these are NOT supposed to be
exposed to the average user, but it made writing the tests more
straightforward.