Removed networking, started boost implementation
I decided that the low-level details of network programming was beyond
the scope of this project. While that is an area I am interested in, I
should write that library separately and then write a back-end for this
project that uses it rather than stuffing a networking library into this
project.