moved to standalone asio,free functions simplified
I was running into linker issues when trying to compile the Boost asio
library into my shared object. I ended up adding the standalone asio
repository as a submodule of this project and is compiled at the same
time as the rest of the code.

The free functions for entrance requests and responses now take a Type *
instead of a Type **. They took a double-ponter before so that they
could set the pointer to NULL to help avoid dangling pointers. However,
this breaks the single responsibility principle, and it burned me when I
started implementing the code to translate the c functions into guile
functions.