Reorganized file hierarchy, refactored code
File hierarchy was organized so that the source folder can be added to
the load path of guile for automatic loading of modules. I am also using
the extension hpp for cpp header files for ease of use with
deoplete-clang.

Refactoring touched butler.guile and attendant.hpp. Code was moved out
of butler.guile and into attendant.guile and glue.guile; additionally, I
created a log module that butler now uses.

The only change to attendant is that Data are now passed around by
pointer instead of by copy.